+++ src/auto_helper_functions_mod.rs
@@ -47,7 +47,7 @@ pub fn run_shell_command_output(shell_command: &str) -> ShellOutput {
 /// Run one shell command and return true if success
 pub fn run_shell_command_success(shell_command: &str) -> bool {
     if !shell_command.starts_with("echo ") && !shell_command.starts_with("printf ") {
-        println!("    {YELLOW}$ {shell_command}{RESET}",);
+        println!("    {YELLOW}$ {shell_command}{RESET}");
     }
     let status = std::process::Command::new("sh").arg("-c").arg(shell_command).status().unwrap();
     // return